+++ src/test/groovy/ru/vyarus/dropwizard/guice/diagnostic/BaseDiagnosticTest.groovy
@@ -2,9 +2,9 @@ package ru.vyarus.dropwizard.guice.diagnostic
 
 import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.bundle.lookup.PropertyBundleLookup
-import ru.vyarus.dropwizard.guice.injector.lookup.InjectorLookup
-import ru.vyarus.dropwizard.guice.module.jersey.debug.HK2DebugBundle
 import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook
+import ru.vyarus.dropwizard.guice.module.context.SharedConfigurationState
+import ru.vyarus.dropwizard.guice.module.jersey.debug.HK2DebugBundle
 import ru.vyarus.dropwizard.guice.support.util.GuiceRestrictedConfigBundle
 import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyHooks
 import spock.lang.Specification
@@ -22,7 +22,7 @@ abstract class BaseDiagnosticTest extends Specification {
 
     void cleanupSpec() {
         // some tests are intentionally failing so be sure to remove stale applications
-        InjectorLookup.clear()
+        SharedConfigurationState.clear()
     }
 
     static class LookupHook implements GuiceyConfigurationHook {